# sql/plan: more efficient implementation of GroupBy