Fixes #331 

Before:

```
go test -benchmem -run=^$ gopkg.in/src-d/go-mysql-server.v0/sql/plan -bench ^BenchmarkGroupBy$

goos: darwin
goarch: amd64
pkg: gopkg.in/src-d/go-mysql-server.v0/sql/plan
BenchmarkGroupBy/no_grouping-4         	     500	   3735067 ns/op	 2468276 B/op	   10075 allocs/op
BenchmarkGroupBy/grouping-4            	     200	   6032010 ns/op	 2386503 B/op	   41453 allocs/op
PASS
ok  	gopkg.in/src-d/go-mysql-server.v0/sql/plan	4.124s
```

Now:

```
go test -benchmem -run=^$ gopkg.in/src-d/go-mysql-server.v0/sql/plan -bench ^BenchmarkGroupBy$

goos: darwin
goarch: amd64
pkg: gopkg.in/src-d/go-mysql-server.v0/sql/plan
BenchmarkGroupBy/no_grouping-4         	    1000	   1641740 ns/op	  321972 B/op	   10033 allocs/op
BenchmarkGroupBy/grouping-4            	     300	   4580011 ns/op	  677216 B/op	   40300 allocs/op
PASS
ok  	gopkg.in/src-d/go-mysql-server.v0/sql/plan	3.706s
```

Aggregations without a grouping clause are now more than 2x faster, 25-30% faster with grouping. Memory usage is an order of magnitude lower than before.

Even if having the code branching into grouping/no grouping and end up with some repeated code, I preferred that path because, as the benchmarks say, it's way faster and it's a pretty common use case to aggregate without grouping.